make the highest kx continuous with next to highest kx. This change is cosmetic but useful for debugging, plots, etc

+++ tests/test_fsaw_noise.py
@@ -9,7 +9,7 @@ def test_wav_admissibility():
                                 nforw=[0, 12, 12, 12, 12, 24, 24, 24, 24],
                                 nback=[18],
                                 pforw=[0, 12, 9, 6, 3, 24, 18, 12, 6],
-                                dtype=float)
+                                dtype=np.float32)
     
     a = np.zeros((shape[-2], shape[-1]//2+1), dtype=fk._cdtype)
     for kern in fk.kernels.values():
@@ -26,7 +26,7 @@ def test_wav_reconstruction():
                                 nforw=[0, 12, 12, 12, 12, 24, 24, 24, 24],
                                 nback=[18],
                                 pforw=[0, 12, 9, 6, 3, 24, 18, 12, 6],
-                                dtype=float)
+                                dtype=np.float32)
 
     rng = np.random.default_rng(0)
     a = rng.standard_normal(shape, dtype=np.float32)
